fetch: split resolution dispatch into a peer module
